refactor: StaffNoteLine::setVolume takes an int
Matches the field's actual type and avoids many unnecessary conversions
between double and int. There's actually only one use of the method with
a double.

+++ src/backend/saving/smp/SMPDecoder.java
@@ -82,7 +82,7 @@ public class SMPDecoder implements Decoder<StaffSequence> {
 								+ Integer.parseInt(meas[1]);
 					} else {
 						if (spl.contains("VOL")) {
-							st.setVolume(Double.parseDouble(spl.substring(
+							st.setVolume((int) Double.parseDouble(spl.substring(
 									spl.indexOf(":") + 1).trim()));
 						} else {
 							try {
